[refactor] #324: 불필요한 VolumeSliderView를 삭제하고 MPVolumeView를 사용하기 쉽게 수정
